+++ awscli/customizations/s3/executor.py
@@ -274,15 +274,15 @@ class PrintThread(threading.Thread):
 
     def _process_print_task(self, print_task):
         print_str = print_task.message
+        print_to_stderr = False
         if print_task.error:
             self.num_errors_seen += 1
-        warning = False
-        if print_task.warning:
-            if print_task.warning:
-                warning = True
-                self.num_warnings_seen += 1
+            print_to_stderr = True
+
         final_str = ''
-        if warning:
+        if print_task.warning:
+            self.num_warnings_seen += 1
+            print_to_stderr = True
             final_str += print_str.ljust(self._progress_length, ' ')
             final_str += '\n'
         elif print_task.total_parts:
@@ -309,21 +309,30 @@ class PrintThread(threading.Thread):
                 self._num_parts += 1
             self._file_count += 1
 
+        # If the message is an error or warning, print it to standard error.
+        if print_to_stderr and not self._quiet:
+            uni_print(final_str, sys.stderr)
+            final_str = ''
+
         is_done = self._total_files == self._file_count
         if not is_done:
-            prog_str = "Completed %s " % self._num_parts
-            num_files = self._total_files
-            if self._total_files != '...':
-                prog_str += "of %s " % self._total_parts
-                num_files = self._total_files - self._file_count
-            prog_str += "part(s) with %s file(s) remaining" % \
-                num_files
-            length_prog = len(prog_str)
-            prog_str += '\r'
-            prog_str = prog_str.ljust(self._progress_length, ' ')
-            self._progress_length = length_prog
-            final_str += prog_str
+            final_str += self._make_progress_bar()
         if not self._quiet:
             uni_print(final_str)
             self._needs_newline = not final_str.endswith('\n')
-            sys.stdout.flush()
+
+    def _make_progress_bar(self):
+        """Creates the progress bar string to print out."""
+
+        prog_str = "Completed %s " % self._num_parts
+        num_files = self._total_files
+        if self._total_files != '...':
+            prog_str += "of %s " % self._total_parts
+            num_files = self._total_files - self._file_count
+        prog_str += "part(s) with %s file(s) remaining" % \
+            num_files
+        length_prog = len(prog_str)
+        prog_str += '\r'
+        prog_str = prog_str.ljust(self._progress_length, ' ')
+        self._progress_length = length_prog
+        return prog_str